[#3578794]Quickfix for history dropdown menu.
The history drop-down menu is bound to 'Next' button and since the commit
91982db0dca1c6e714d6db4ea068e648a28082eb this item isn't available if there
is no forward history. Since this case is often enough and as soon there is
backward history the drop-down menu should be available the menu moved to
'Prev' button instead.
The side effect of that is drop-down menu still isn't available if there is
no backward history but there is some forward one. That case is very rare
and making it available requires big redesign. It's why this change is marked
as Quickfix. If mentioned case will be ever requested it can be feature req.