Added wallpaper mode 'crop'.
Hi there,

I have added a (maybe even "the") missing wallpaper mode to pcmanfm.

I can be described as follows: The image will be cropped by the maximum
necessary amount to fill the whole screen without changing the aspect ratio
(i.e. stretch).

Technically it is a variation of the 'fit' mode. The difference is that it
uses the greater of the width/height aspect ratios to scale the image.
Thus it cuts off the image on two sides, as compaired to filling (the other)
two sides with the background color.

I have not added any translations yet. I'm not sure what the workflow would
be here. One combo box item was added: "Crop to fill the entire screen".
The german translation (considering the existing translations of the other
items) would be "Zuschneiden".

Regards
Jonas Petersen

Signed-off-by: Jonas Petersen <jnsptrsn1@gmail.com>