[#3591767]Don't hilight desktop item if we cannot drop on it.
If we hilight every item on desktop then we step into a problem with
desktop item moving - it hilights itself but if we don't actually move
the item then it stays hilighted because we don't perform any action
in that case but we cannot get any notification on drop (just because
there was no action performed) therefore we cannot clear hilight.

Since not hilighting desktop item which we cannot drop onto looks
reasonable as well, we just disable it and that solves the problem.