Fix handling extended sort modes of folder view.
When we change extended modes such as FM_SORT_CASE_SENSITIVE, the FmMainWin
does not handle it if no main menu update was made (appropriate items aren't
present in it yet). So on sorting change we should check it and save the
change into main config if some of extended modes was changed.