Save per-folder config: sort mode, show_hidden, and view mode.
The settings will be loaded and saved in the file .directory if it exists,
otherwise the file ~/.config/pcmanfm/<profile>/dir-settings.conf will be
used to save all settings.

The settings will be saved only if 'View'->'Save for This Folder' checkbox
in the main menu is checked, and will be forgotten when it's unchecked. No
big volume of data is expected therefore.

The format of the file is the same as format of global settings for the
same data (pcmanfm.conf) for few obvious reasons:
- not add another format, all existing data are in key file format;
- not add additional dependencies for pcmanfm.
Data are cached in pcmanfm so no impact on performance is expected.