Correctly restore the scroll position after navigating history.
When we go to another folder then come back and try to restore scroller
it's done only in range of first page. That is before the view isn't
yet rendered at the moment of finish of folder loading. To fix that
problem we should set scroll position after it is rendered fully.