Implemented handling for extra items (Trash, Documents, media) on the desktop.
Those items have not be in the Desktop folder, they will be added to
the view by desktop manager internally. Corresponding config options
are 'show_trash', 'show_documents', 'show_mounts' - those are set in
the desktop config for each monitor. Default is true for 'show_trash'
and false for other items. Note that all those items are not links of
any kind but folders itself so weird operations with them may lead to
weird results.