[#3613571]Check if desktop background file was changed and reset if it was.
The check in the commit is done by comparing file modification time.
This should cover almost all the cases. Yes, it adds some I/O on each
desktop switching to retrieve mtime but in most of cases that data is
cached by the kernel so that should not create any overheads on I/O.