[#1434810]new menu option 'View'->'Filter...' to set filter for files.
The option opens a simple dialog to enter a shell pattern to filter files
currently shown in the view. The pattern is case insensitive and applied
to displayed names in the view. The pattern is persistent for the tab page
until user replaces it with another pattern ("*" for show all).