Add an option to open files with a single middle-click
With this option, a single middle-click has the same effect as a
double-click with the left mouse button.

The single left-click and the single middle-click options are mutually
exclusive, i.e. both options cannot be selected at the same time, so
activating one option automatically deactivates the other.

This uses the middle mouse button differently to the pull request
lxde/libfm#46 by livanh.